feat: 阶段1 S1.1-S1.5 —— standalone 重构 + Flight SQL 标准协议（架构 v12.3）
- S1.1 结构调整：撤销 bins/，all-in-one → crates/yuntun-standalone（bin 名 yuntun）
- S1.2-S1.5 Flight SQL 标准协议：语句查询 / 元数据（Catalogs/Tables 含 schema）/
  prepared statement / 批量装载（StatementIngest + prepared INSERT 绑定数据）
- 架构简化（v12.3）：协议端口统一在 server 节点层，FlightServer 直接组合
  Arc<Ingestor> + Arc<QueryEngine>；删除 Hook/Gateway 间接层；
  ingest/query 回归纯能力 crate（移除协议依赖）；所有写入走 ingest 管线（唯一写入事实）
- 修复 Windows 目录 fsync（FILE_FLAG_BACKUP_SEMANTICS），WAL/chaos 测试在 Windows 首次全绿
- 修复 4 个 ADBC Go 驱动互操作问题（action Any 包装 / schema IPC 封装格式 /
  省略 endpoint location / 元数据 schema 对齐官方）
- QueryEngine 开启 information_schema 并新增 schema_of()（逻辑计划取 schema）
- 新增 scripts/pyarrow_smoke.py（ADBC + 手写 wire 冒烟）；docs 四份文档同步 v12.3
- 全量测试绿、clippy --workspace --all-targets 零警告、cargo fmt 已格式化

+++ crates/ingest/src/accumulator.rs
@@ -7,9 +7,9 @@
 //! - 绝对空闲超时兜底：5 分钟（防定时器 bug 导致无限滞留）
 
 use crate::timeutil::{jitter_seconds, window_start_ms, MINUTE_MS};
-use yuntun_model::wal_record::DataPayload;
 use std::collections::HashMap;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use yuntun_model::wal_record::DataPayload;
 
 /// 窗口归属：数据按 event_time（缺省用 received_at）对齐到整分钟（ADR-10）。
 pub fn window_of(event_time_ms: Option<i64>, received_at_ms: i64) -> (i64, String) {
@@ -24,10 +24,18 @@ pub fn extract_event_time_ms(batch: &arrow::record_batch::RecordBatch) -> Option
     let arr = batch.column(idx);
     use arrow::array::{Array, Int64Array, TimestampMillisecondArray};
     if let Some(a) = arr.as_any().downcast_ref::<Int64Array>() {
-        return if a.null_count() == arr.len() { None } else { Some(a.value(0)) };
+        return if a.null_count() == arr.len() {
+            None
+        } else {
+            Some(a.value(0))
+        };
     }
     if let Some(a) = arr.as_any().downcast_ref::<TimestampMillisecondArray>() {
-        return if a.null_count() == arr.len() { None } else { Some(a.value(0)) };
+        return if a.null_count() == arr.len() {
+            None
+        } else {
+            Some(a.value(0))
+        };
     }
     None
 }
@@ -53,7 +61,7 @@ impl WindowGroup {
     /// flush 触发判定（§5.3 任一满足即 flush）。
     pub fn should_flush(&self, now_ms: u64, cfg: &crate::pipeline::IngestorConfig) -> bool {
         // ① 行数阈值
-        if self.rows >= cfg.rows_threshold as u64 {
+        if self.rows >= cfg.rows_threshold {
             return true;
         }
         // ② 时间阈值：距窗口开始 >= time_threshold
@@ -207,7 +215,10 @@ mod tests {
             },
         );
         assert!(acc.drain_ready(now, &cfg()).is_empty());
-        let g = acc.groups.get_mut(&("t".into(), "s0".into(), "w1".into())).unwrap();
+        let g = acc
+            .groups
+            .get_mut(&("t".into(), "s0".into(), "w1".into()))
+            .unwrap();
         g.rows += 1;
         let ready = acc.drain_ready(now, &cfg());
         assert_eq!(ready.len(), 1);
